Fix: using Control Center invalidate Touch ID unlock.
Before fix:
ANYTIME when Control Center is dismissed, resetBiometricMatchingEnabled() would be called, which make Touch ID recognise tapping quicker but invalidate identity matching. (This is a fix for BioLockdown. See comments in source code.)
But it should not be called when device is locked, because identity matching is required for Touch ID unlock.

This fix checks if device is locked before calling resetBiometricMatchingEnabled().